make steal-ejs work again
